requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
clize==5.0.2
datasets==2.18.0
dill==0.3.8
docutils==0.21
fasttext==0.9.2
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.3
joblib==1.4.0
lxml==5.2.1
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numlet==2.0.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
od==2.0.2
packaging==24.0
pandas==2.2.1
Paquete-nlt-roylanmartinez @ git+https://github.com/roylanmartinez/Numlet.git@fb7f8ab294dbc4861f9f264d79da054f7c13ac19
pillow==10.3.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pybind11==2.12.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sympy==1.12
threadpoolctl==3.4.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.2
transformers==4.40.0
triton==2.2.0
typing_extensions==4.11.0
tzdata==2024.1
unicode==2.9
Unidecode==1.3.8
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
